+++ src/pageEditor/fields/SelectorSelectorWidget.tsx
@@ -36,10 +36,8 @@ import {
   cancelSelect,
 } from "@/contentScript/messenger/api";
 import { thisTab } from "@/pageEditor/utils";
-import {
-  type ElementInfo,
-  type SelectMode,
-} from "@/contentScript/pageEditor/types";
+import { type SelectMode } from "@/contentScript/pageEditor/types";
+import { type ElementInfo } from "@/pageScript/frameworks";
 import { useSelector } from "react-redux";
 import { type SettingsState } from "@/store/settingsTypes";
 import { sortBySelector } from "@/utils/inference/selectorInference";
