+++ src/pageEditor/sidebar/SidebarExpanded.tsx
@@ -18,7 +18,7 @@
 import styles from "./Sidebar.module.scss";
 import React, { type FormEvent, useEffect, useMemo, useState } from "react";
 import { sortBy } from "lodash";
-import { getRecipeById } from "@/utils";
+import { getRecipeById } from "@/pageEditor/utils";
 import { Accordion, Button, Form, ListGroup } from "react-bootstrap";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import hash from "object-hash";
