+++ src/pageEditor/utils.ts
@@ -38,6 +38,7 @@ import { PIPELINE_BLOCKS_FIELD_NAME } from "./consts";
 import { isExpression } from "@/runtime/mapArgs";
 import { expectContext } from "@/utils/expectContext";
 import DisplayTemporaryInfo from "@/blocks/transformers/temporaryInfo/DisplayTemporaryInfo";
+import { RecipeDefinition } from "@/types/definitions";
 
 export async function getCurrentURL(): Promise<string> {
   expectContext("devTools");
@@ -66,6 +67,13 @@ export function getRecipeIdForElement(
   element: IExtension | FormState
 ): RegistryId {
   return isExtension(element) ? element._recipe?.id : element.recipe?.id;
+}
+
+export function getRecipeById(
+  recipes: RecipeDefinition[],
+  id: RegistryId
+): RecipeDefinition | undefined {
+  return recipes.find((recipe) => recipe.metadata.id === id);
 }
 
 export function getPipelinePropNames(block: BlockConfig): string[] {
