+++ src/pageScript/frameworks.ts
@@ -19,8 +19,15 @@ import { type Framework } from "@/messaging/constants";
 import adapters from "@/pageScript/frameworks/adapters";
 import { isEmpty, uniq } from "lodash";
 import { inferSelectorsIncludingStableAncestors } from "@/utils/inference/selectorInference";
-import { type ElementInfo } from "@/contentScript/pageEditor/types";
 
+export interface ElementInfo {
+  selectors: string[];
+  framework: Framework;
+  tagName: string;
+  hasData: boolean;
+  parent?: ElementInfo;
+  isMulti?: boolean;
+}
 export async function elementInfo(
   element: HTMLElement,
   componentFramework?: Framework,
